Put constants together and change if
While importing from the same module, it is a good idea to use one import statement - doing it separately provides unnecessary boilerplate code.

When it comes to if statement, empty strings/empty lists etc. are evaluated to False in python and can be treated as such. Hence, if not user_choice.

Also, it is risky to catch exceptions like this (ie with Exception). It literally catches all exceptions.
For instance, the following will not crush the program:
try:
    something
except Exception as e:
    passs
Therefore, it is better to catch individual exceptions like ValueError because it is less ambiguous.
I've no experience with Kivy though, so I can't say what should be caught in particular and this is only a suggestion.